Add test for kmeans_n_times_csv

+++ multi/kmeans.py
@@ -47,7 +47,7 @@ def kmeans_n_times_csv(X, n, num_clusters, csv_file, newcsv=True, **kwargs):
       
     
     #Make empty dataframe
-    df_kmeans = pd.DataFrame(index=[f'sample_{i}' for i in range(len(X))])
+    df_kmeans = pd.DataFrame(index=[f'sample_{i}' for i in range(len(X))],dtype='int32')
     
     #Add the corresponding labels
     if 'labels' in kwargs:
